Added global rankings, controllers updated.

+++ src/main/java/com/score_keeper/controllers/StageController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
@@ -19,10 +20,9 @@ public class StageController {
     StageRepository stageRepository;
     @Autowired
     TournamentRepository tournamentRepository;
-    @Autowired
-    ClubRepository clubRepository;
 
-    @GetMapping(value = {"/", ""})
+
+    @GetMapping(value = "/stages")
     public Map<String, Object> getStages() {
         HashMap<String, Object> response = new HashMap<>();
         try {
@@ -38,33 +38,39 @@ public class StageController {
 
     }
 
-    @PostMapping(value = {"/", ""})
-    public Map<String, Object> postPlayer(@Valid @RequestBody Stage data) {
-        HashMap<String, Object> response = new HashMap<>();
-        try {
-            Tournament tournament = tournamentRepository.findById(data.getTournament().getId()).get();
-            if (tournament.isBlocked() || data.isBlocked()) {
-                response.put("message", "Stage is blocked");
-                response.put("success", true);
-                return response;
-            }
-            data.setTournament(tournament);
-            Club club = clubRepository.findById(data.getClub().getId()).get();
-            data.setClub(club);
-            stageRepository.save(data);
-            response.put("stage", data);
-            response.put("message", "Successful");
-            response.put("success", true);
-            return response;
-        } catch (Exception e) {
-            response.put("message", e.getMessage());
-            response.put("success", false);
-            return response;
-        }
+    //Never used
+//    @PostMapping(value = "/stages")
+//    public Map<String, Object> postPlayer(@Valid @RequestBody Stage data) {
+//        HashMap<String, Object> response = new HashMap<>();
+//        try {
+//            Tournament tournament = tournamentRepository.findById(data.getTournament().getId()).get();
+//            if (tournament.isBlocked() || data.isBlocked()) {
+//                response.put("message", "Stage is blocked");
+//                response.put("success", true);
+//                return response;
+//            }
+//            if(tournament.getStages() < data.getStageNumber()){
+//                response.put("message", "Stage number surpass tournament stages");
+//                response.put("success", true);
+//                return response;
+//            }
+//            data.setTournament(tournament);
+//
+//            stageRepository.save(data);
+//            response.put("stage", data);
+//            response.put("message", "Successful");
+//            response.put("success", true);
+//            return response;
+//        } catch (Exception e) {
+//            response.put("message", e.getMessage());
+//            response.put("success", false);
+//            return response;
+//        }
+//
+//    }
 
-    }
 
-    @GetMapping(value = "/{id}")
+    @GetMapping(value = "/stages/{id}")
     public Map<String, Object> getStageById(@PathVariable("id") String id) {
         HashMap<String, Object> response = new HashMap<String, Object>();
         try {
@@ -85,12 +91,46 @@ public class StageController {
         }
     }
 
+//    @GetMapping(value = "/tournament/{tournamentID}/stages")
+//    public Map<String, Object> getStagesByTournamentId(@PathVariable("tournamentID") String id) {
+//        HashMap<String, Object> response = new HashMap<String, Object>();
+//        try {
+//            List<Stage> stages = stageRepository.findAllByTournamentId(id);
+//            if (!stages.isEmpty()) {
+//                response.put("data", stages);
+//                response.put("message", "Stages found");
+//                response.put("success", true);
+//            } else {
+//                response.put("message", "Stages not found with  tournament id " + id);
+//                response.put("success", false);
+//            }
+//            return response;
+//        } catch (Exception e) {
+//            response.put("message", "" + e.getMessage());
+//            response.put("success", false);
+//            return response;
+//        }
+//    }
+
     @PutMapping(value = "/{id}")
     public Map<String, Object> updateStage(@PathVariable("id") String id, @Valid @RequestBody Stage stage) {
         HashMap<String, Object> response = new HashMap<String, Object>();
         try {
-            if (stageRepository.findById(id).get().getTournament().isBlocked() || stageRepository.findById(id).get().isBlocked()) {
+
+            if(stageRepository.findById(id).isPresent()){
+                response.put("message", "Stage not found with id " + id);
+                response.put("success", false);
+                return response;
+            }
+            Tournament tournament = tournamentRepository.findById(stage.getTournament().getId()).get();
+
+            if (tournament.isBlocked() || stageRepository.findById(id).get().isBlocked()) {
                 response.put("message", "Stage is blocked");
+                response.put("success", true);
+                return response;
+            }
+            if(tournament.getStages() < stage.getStageNumber()){
+                response.put("message", "Stage number surpass tournament stages");
                 response.put("success", true);
                 return response;
             }
@@ -102,12 +142,10 @@ public class StageController {
             }
             if (stageOptional.isPresent()) {
                 stage.setId(id);
+                stage.setTournament(stage.getTournament());
                 stageRepository.save(stage);
-                response.put("message", "Successfully update");
+                response.put("message", "Successfully updated");
                 response.put("success", true);
-            } else {
-                response.put("message", "Stage not found with id " + id);
-                response.put("success", false);
             }
             return response;
         } catch (Exception e) {
@@ -121,6 +159,11 @@ public class StageController {
     public Map<String, Object> deleteStage(@PathVariable("id") String id) {
         HashMap<String, Object> response = new HashMap<String, Object>();
         try {
+            if (stageRepository.findById(id).get().getTournament().isBlocked() || stageRepository.findById(id).get().isBlocked()) {
+                response.put("message", "Stage is blocked");
+                response.put("success", true);
+                return response;
+            }
             Optional<Stage> stageOptional = stageRepository.findById(id);
             if (stageOptional.isPresent()) {
                 stageRepository.deleteById(id);
@@ -138,7 +181,7 @@ public class StageController {
         }
     }
 
-    @PostMapping(value = "/block/{id}")
+    @PutMapping(value = "/block/{id}")
     public Map<String, Object> blockStage(@PathVariable("id") String id){
         HashMap<String, Object> response = new HashMap<String, Object>();
         try {
@@ -151,7 +194,7 @@ public class StageController {
                 response.put("message", "Stage blocked");
                 response.put("success", true);
             } else {
-                response.put("message", "Tournament not found with id " + id);
+                response.put("message", "Stage not found with id " + id);
                 response.put("success", false);
             }
             return response;
@@ -162,7 +205,7 @@ public class StageController {
         }
     }
 
-    @PostMapping(value = "/unblock/{id}")
+    @PutMapping(value = "/unblock/{id}")
     public Map<String, Object> unblockStage(@PathVariable("id") String id){
         HashMap<String, Object> response = new HashMap<String, Object>();
         try {
